libs/framework/eventservice/src/events/shoppingBag/addEvoucher/index.test.ts
/* eslint-disable @typescript-eslint/unbound-method */
import {AddEVoucher, AddEVoucherCallback} from "."
import Emitter from "../../../eventEmitter"

jest.mock("../../../eventEmitter", () => ({
    emit: jest.fn(),
    listen: jest.fn(),
}))

describe("Given a AddEVoucher ESB", () => {
    describe("When publishing", () => {
        const mockPublishData = {
            id: "1",
            option: "2",
            eVoucherTo: "3",
            eVoucherFrom: "4",
            eVoucherEmail: "5",
            eVoucherMessage: "6",
            ecardDeliveryDate: "7",
        }

        beforeAll(() => {
            const getBackEsb = new AddEVoucher()
            getBackEsb.publish(mockPublishData)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.emit).toHaveBeenCalledWith("SHOPPING_BAG_ADD_EVOUCHER", mockPublishData)
        })
    })

    describe("When subscribing", () => {
        const mockCallback = jest.fn()

        beforeAll(() => {
            const getBackEsb = new AddEVoucher()
            getBackEsb.subscribe(mockCallback)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.listen).toHaveBeenCalledWith("SHOPPING_BAG_ADD_EVOUCHER", mockCallback)
        })
    })
})

describe("Given a AddEVoucherCallback ESB", () => {
    describe("When publishing", () => {
        const mockPublishData = {
            success: true,
            data: {},
            status: 123,
            textStatus: "test",
        }

        beforeAll(() => {
            const getBagCallbackEsb = new AddEVoucherCallback()
            getBagCallbackEsb.publish(mockPublishData)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.emit).toHaveBeenCalledWith("SHOPPING_BAG_ADD_EVOUCHER_CALLBACK", mockPublishData)
        })
    })

    describe("When subscribing", () => {
        const mockCallback = jest.fn()

        beforeAll(() => {
            const getBagCallbackEsb = new AddEVoucherCallback()
            getBagCallbackEsb.subscribe(mockCallback)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.listen).toHaveBeenCalledWith("SHOPPING_BAG_ADD_EVOUCHER_CALLBACK", mockCallback)
        })
    })
})
